Skip invalid pear channels when building the composer index
For older packages, there are cases where dependencies no longer exist on a channel. This is especially true for phpunit, whose pear channel no longer exists.

Rather than break the entire build, we WARN that this is happening and allow pear channel maintainers to make a judgement call as to whether the composer repo is important (in this case, likely not).